+++ qpsolvers/solvers/qpswift_.py
@@ -30,12 +30,140 @@ If you use qpSWIFT in your research, consider citing the corresponding paper
 [Pandala2019]_.
 """
 
+import warnings
 from typing import Optional
 
 import numpy as np
 import qpSWIFT
 
-from ..conversions import linear_from_box_inequalities
+from ..conversions import linear_from_box_inequalities, split_dual_linear_box
+from ..problem import Problem
+from ..solution import Solution
+
+
+def qpswift_solve_problem(
+    problem: Problem,
+    initvals: Optional[np.ndarray] = None,
+    verbose: bool = False,
+    **kwargs,
+) -> Solution:
+    """
+    Solve a quadratic program using `qpSWIFT
+    <https://github.com/qpSWIFT/qpSWIFT>`__.
+
+    Note
+    ----
+    This solver does not handle problems without inequality constraints.
+
+    Parameters
+    ----------
+    problem :
+        Quadratic program to solve.
+    initvals :
+        Warm-start guess vector.
+    verbose :
+        Set to `True` to print out extra information.
+
+    Returns
+    -------
+    :
+        Solution returned by the solver.
+
+    Note
+    ----
+    .. _qpSWIFT rank assumptions:
+
+    **Rank assumptions:** qpSWIFT requires the QP matrices to satisfy the
+
+    .. math::
+
+        \\begin{split}\\begin{array}{cc}
+        \\mathrm{rank}(A) = p
+        &
+        \\mathrm{rank}([P\\ A^T\\ G^T]) = n
+        \\end{array}\\end{split}
+
+    where :math:`p` is the number of rows of :math:`A` and :math:`n` is the
+    number of optimization variables. This is the same requirement as
+    :func:`cvxopt_solve_qp`, however qpSWIFT does not perform rank checks as it
+    prioritizes performance. If the solver fails on your problem, try running
+    CVXOPT on it for rank checks.
+
+    Notes
+    -----
+    All other keyword arguments are forwarded as options to the qpSWIFT solver.
+    For instance, you can call ``qpswift_solve_qp(P, q, G, h, ABSTOL=1e-5)``.
+    See the solver documentation for details.
+
+    For a quick overview, the solver accepts the following settings:
+
+    .. list-table::
+       :widths: 30 70
+       :header-rows: 1
+
+       * - Name
+         - Effect
+       * - MAXITER
+         - Maximum number of iterations needed.
+       * - ABSTOL
+         - Absolute tolerance on the duality gap. See *e.g.* [tolerances]_ for
+           a primer on the duality gap and solver tolerances.
+       * - RELTOL
+         - Relative tolerance on the residuals :math:`r_x = P x + G^T z + q`
+           (dual residual), :math:`r_y = A x - b` (primal residual on equality
+           constraints) and :math:`r_z = h - G x - s` (primal residual on
+           inequality constraints). See equation (21) in [Pandala2019]_.
+       * - SIGMA
+         - Maximum centering allowed.
+
+    If a verbose output shows that the maximum number of iterations is reached,
+    check e.g. (1) the rank of your equality constraint matrix and (2) that
+    your inequality constraint matrix does not have zero rows.
+
+    As qpSWIFT does not sanity check its inputs, it should be used with a
+    little more care than the other solvers. For instance, make sure you don't
+    have zero rows in your input matrices, as it can `make the solver
+    numerically unstable <https://github.com/qpSWIFT/qpSWIFT/issues/3>`_.
+    """
+    if initvals is not None:
+        print("qpSWIFT: warm-start values are ignored by wrapper")
+    P, q, G, h, A, b, lb, ub = problem.unpack()
+    if lb is not None or ub is not None:
+        G, h = linear_from_box_inequalities(G, h, lb, ub)
+    result: dict = {}
+    kwargs.update(
+        {
+            "OUTPUT": 2,  # include "sol", "basicInfo" and "advInfo"
+            "VERBOSE": 1 if verbose else 0,
+        }
+    )
+    if G is not None and h is not None:
+        if A is not None and b is not None:
+            result = qpSWIFT.run(q, h, P, G, A, b, kwargs)
+        else:  # no equality constraint
+            result = qpSWIFT.run(q, h, P, G, opts=kwargs)
+    else:  # no inequality constraint
+        # See https://github.com/qpSWIFT/qpSWIFT/issues/2
+        raise NotImplementedError(
+            "QPs without inequality constraints are not handled by qpSWIFT"
+        )
+    basic_info = result["basicInfo"]
+    adv_info = result["advInfo"]
+
+    solution = Solution(problem)
+    solution.extras.update(basic_info)
+    solution.extras.update(adv_info)
+    solution.obj = adv_info["fval"]
+    exit_flag = basic_info["ExitFlag"]
+    if exit_flag != 0:
+        return solution
+    solution.x = result["sol"]
+    if A is not None:
+        solution.y = adv_info["y"]
+    z, z_box = split_dual_linear_box(adv_info["z"], P.shape[0], lb, ub)
+    solution.z = z
+    solution.z_box = z_box
+    return solution
 
 
 def qpswift_solve_qp(
@@ -162,28 +290,12 @@ def qpswift_solve_qp(
     have zero rows in your input matrices, as it can `make the solver
     numerically unstable <https://github.com/qpSWIFT/qpSWIFT/issues/3>`_.
     """
-    if initvals is not None:
-        print("qpSWIFT: warm-start values are ignored by wrapper")
-    if lb is not None or ub is not None:
-        G, h = linear_from_box_inequalities(G, h, lb, ub)
-    result: dict = {}
-    kwargs.update(
-        {
-            "OUTPUT": 1,  # include "sol" and "basicInfo"
-            "VERBOSE": 1 if verbose else 0,
-        }
+    warnings.warn(
+        "The return type of this function will change "
+        "to qpsolvers.Solution in qpsolvers v3.0",
+        DeprecationWarning,
+        stacklevel=2,
     )
-    if G is not None and h is not None:
-        if A is not None and b is not None:
-            result = qpSWIFT.run(q, h, P, G, A, b, kwargs)
-        else:  # no equality constraint
-            result = qpSWIFT.run(q, h, P, G, opts=kwargs)
-    else:  # no inequality constraint
-        # See https://github.com/qpSWIFT/qpSWIFT/issues/2
-        raise NotImplementedError(
-            "QPs without inequality constraints are not handled by qpSWIFT"
-        )
-    exit_flag = result["basicInfo"]["ExitFlag"]
-    if exit_flag != 0:
-        return None
-    return result["sol"]
+    problem = Problem(P, q, G, h, A, b, lb, ub)
+    solution = qpswift_solve_problem(problem, initvals, verbose, **kwargs)
+    return solution.x